# Optimize Integer#& and Integer#|