--- a/vm.c
+++ b/vm.c
@@ -1611,6 +1611,7 @@ vm_init_redefined_flag(void)
     OP(Min, MIN), (C(Array));
     OP(Call, CALL), (C(Proc));
     OP(And, AND), (C(Integer));
+    OP(Or, OR), (C(Integer));
 #undef C
 #undef OP
 }